Add CRUD api to Trainee_Course_Management

+++ src/main/java/com/example/trainingsystem/models/embedded/CompletedClass_Trainee.java
@@ -11,5 +11,5 @@ public class CompletedClass_Trainee {
     private String ClassName;
 
     @JsonProperty("TotalScore")
-    private int TotalScore;
+    private double TotalScore;
 }
